Change assumptions on longitude conventions
Previously, assumed that we were only modeling for Atlantic basin storms, and longitudes were converted to degrees West (positive number for most Atlantic-basin storms) when imputing the storm tracks. We are changing the code to make it applicable in any basin. This change allows storm tracks to use negative values in the Western hemisphere for longitude (i.e., always using degrees east), which seems more consistent with how tracking data are available through IBTrACS. The code has always used this convention for its original inputs, but then was changing conventions within parts of the code---with this change, we're going back to a more universal convention throughout.