Change to use double exponential at all r > R1
Previously, within the transition region, we were using eq 1b from Willoughby, then the double exponential outside R2 (the radius defining the end of the transition region). This changes to use the double exponential from the start of the transition region (R1). Based on equation 4 in Willoughby, the double exponential can be used for any radius of R1 or higher.