Incorporate land/sea of location for wind model
This changes the gradient to surface calculation and the add inflow calculation to incorporate whether the location being modeled is over land or water. It inputs a logical value for whether the point is over land or water and uses that. Before, it assumed that only points over land were being modeled.