Create an example set of Australian storms
This gives us some Southern Hemisphere examples. It also will be a good data set to use to show how to apply the code across multiple storms. These data are from the IBTrACS CSV file for storms since 1980. They were created by filtering that data to a few storms, selecting some of the columns, and changing all the column names to be lowercase.