Change add_forward_speed to return a vector
Previously, it only returned the windspeed. To get the surface wind direction with the forward speed added in, we need to get two values back from this function, the wind speed and the wind direction once the forward speed of the storm is incorporated. This change is the first step, to make it so that the value returened from `add_forward_speed` has two components, rather than one (using a NumericVector for this). The next step will be to correctly calculate that angle (now it's just using 90 as a placeholder).